[iOS] Use new async clipboard API for UIMenuController
This avoids actually checking the clipboard contents until the user
requests it, preventing the iOS 14 clipboard access notification from
showing unnecessarily.

We request the clipboard state when the corresponding view appears and
whenever the clipboard changes again. This has the small issue that the
edit menu may be incorrect if it is presented before the asynchronous
call finishes, but my testing shows that the asnychronous call happens
pretty quickly.

Also, because the omnibox view controller is always a child view
controller of the location bar view controller, the location bar view
controller is in the responder chain, so it can handle the pasting from
both classes.

Bug: 1110848
Change-Id: I1b20f8099abc20408321c602976f9adb6434ba04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2325733
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#794986}