[Code Health] Convert Bind/Callback to Once/Repeating.
`ImageDecoder::StartWithOptionsImpl`, binds to the image decode
success and error callback with `BindOnce`.  This means that the
`ClipboardImageDataDecoder` methods `OnImageDecoded` and
`OnImageFailed` will also only by called once, and so the
`ClipboardExtensionHelper` can therefore store the callbacks as
`OnceCallback`s.

All other changes are purely mechanical.  For example, the method
`CountryComboboxModel::SetCountries` accepts a `RepeatingCallback`.
Previously the callback passed in `GenerateCountryList` was
implicitly converted to a `RepeatingCallback` from a `Callback`.
This makes it explicit.

Bug: 1007786
Change-Id: I5b09978da57ea1df2ff1e672a5db735ab677e965
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2321541
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Commit-Queue: Terrence Reilly <treilly@google.com>
Cr-Commit-Position: refs/heads/master@{#795002}