[FragmentItem] Block fragmentation support in NGInlineCursor
This patch adds block fragmentation functions to
|NGInlineCursor|, as detailed in the design doc:
https://docs.google.com/document/d/1mb_QRS67JQcOZ7oFdDb6UX-gx10IvhLwVPytNLM6wr0/edit?usp=sharing

The implementation is to traverses all fragments, which is
probably not fast enough to ship, but the behavior should
be correct. The performance work will be in future patches.

We can start removing work arounds such as ones in
|NGFragmentItems::LayoutObjectWillBeDestroyed|. Such work
will also be in future patches.

Bug: 829028, 982194, 1104064, 1112657
Change-Id: Ied6fb7b868895669c651418134041a0feda3c0d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2313736
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795011}