Do not start elide animation if there will be no visible change
Currently the ElideAnimation starts even if there are no visible
changes (i.e. the animate from state is equal to the animate to state),
this causes bugs in behaviors that rely on knowing if an animation is
running. This CL adds an early return to ElideAnimation::Start if
animate from and animate to states are the same.

Bug: 1112848
Change-Id: I99966d555c586ba367bfb5ee57429d98d3458318
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335760
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795029}