[Installable] InstallableManager should never return nullptr Manifest.
The new error pathway populated the manifest pointer as a nullptr. The
code can assume that this is never a nullptr, so instead just create
a dummy empty manifest on the stack and send a pointer to that.

Bug: 1110771
Change-Id: If358b4d7251b0a6a858cf91e511dda60ba02ee89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327277
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Auto-Submit: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795037}