Fix URL elision logic with RTL UI
This fixes the URL elision and animation logic when the UI is RTL.
This does not yet address RTL URLs, just fixes the logic when Chrome
itself is in RTL mode, but only works with LTR URLs.
Gradient mask is not currently functional if UI is RTL, so this disables
it in that case.

Bug: 1101472
Change-Id: Ib89ebb6eb21882aa83ec923ec1720f25468bc6d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333293
Auto-Submit: Carlos IL <carlosil@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795060}