Reland "Add UseCounters for Conversion Measurement API"
Reland of https://crrev.com/c/2314843

Change was reverte due to browser test flakes on Linux Asan bots on
https://crrev.com/c/2326796

The browsertest flakes seem to have been the result of a race condition
caused by logging UseCounters while navigating the main frame cross
origin. In these cases, browser-side PLM code received the feature
usage updates after the main frame had already begun navigating, causing
us to ignore the update. This is unexpected, and is summarized on
https://crbug.com/1111524

This change fixes the issue by using an impression tag that targets a
new window, so that we can guarantee the feature is received prior
to flushing metrics for the page with the impression.

TBR=japhet@chromium.org

Bug: 1109079
Change-Id: Ie80a9156fb9e90d9891b02e0bdc44dc6cd0fe266
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330300
Reviewed-by: John Delaney <johnidel@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: John Delaney <johnidel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795067}