[AW][Dev-UI] scrollable warning message in flag UI
This puts the flag UI's warning text into the ListView so it can scroll
off screen as the user scrolls down the list of flags. This adds some
complexity because the ListView now supports two different types of
layouts.

This swaps the order of the warning text and the "reset all" button.
This is in order to pin the "reset all" button to the top of the
Fragment.

Fixed: 1059881
Test: Manual - verify the warning text scrolls as expected
Change-Id: Id7afd2b256051d6d7bf5cee18c73592021444194
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330042
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Hazem Ashmawy <hazems@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795070}