Use native Parent Access Code validation.
When supervised user is the owner of the device, the Parent Access Code
(PAC) is required for performing:
- Add user
- Re-authentication

The check is currently done in Gaia flow. This CL is to migrate the
check to the client using the Parent Access Code widget that is
currently used in couple other places.

Bug: 921142
Test: unit tests
Change-Id: I3588da7779904b7af7145e533af83673eeb8742b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333544
Commit-Queue: Li Lin <llin@chromium.org>
Reviewed-by: Aga Wronska <agawronska@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795079}