Filter RUN-UNTIL-EVENTS by platform
If a single platform requires the behavior of RUN-UNTIL-EVENTS, but the
test is run for multiple platforms, then RUN-UNTIL-EVENTS would need to
be defined for an event per-platform.

This change applies the same concept as filters for WIN-ALLOW/DENY so
we can enable this behavior per-platform for event dump tests.
e.g. WIN-RUN-UNTIL-EVENTS

Bug: none
Change-Id: I0b4eda70c64dea685ed4df8014465297c66a925c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320417
Commit-Queue: Adam Ettenberger <Adam.Ettenberger@microsoft.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Kevin Babbitt <kbabbitt@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#795083}