GMC: Update MediaSessionInfo with audio sink id
The media session controller will now notify the media session of audio
device switching. Upon receiving this message, the media session will
update the MediaSessionInfo to include the device id and notify its
observers.

Because it is possible for a media session's players to be using
different audio devices, a device is only added to the info object if
all of the sessions players are using the same device.

Bug: 1096251
Change-Id: I280387991b333eff60141e38e119eff7444aebb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324406
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Becca Hughes <beccahughes@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795156}