Show token account ID in Identity Internals page
Currently, the identity internals debugging page does not display the
account ID that the token belongs to. This information may be useful to
list for debugging cases where multiple accounts are involved.

This CL adds the account ID for each cached token to the Identity
Internals page.

Bug: None
Change-Id: I704748f2521cb34c96ed7ccade0fdac6e99d01dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333545
Commit-Queue: William Lin <williamlin@google.com>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795162}