Overload GetMediaDeviceIDForHMAC to work with audio output devices.
The current implemnetation of GetMediaDeviceIDFroHMAC accepts a
blink::mojom::MediaStreamType to distinguish between audio and video
devices. However, it is not able to distinguish between audio input and
output. Currently, this method will assume input.

In order to allow this method to work with audio output devices this
change introduces an overload that accepts a blink::MediaDeviceType
instead of a stream type.

Bug: 1111585
Change-Id: Ie7111665f8cd3d24a360bff28d2da56c7ddfaacb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2330741
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: John Rummell <jrummell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795172}