GMC: Allow recipient of OnAudioOutputSinkChanged IPC to translate device id
In order for the global media controls dialog to display the currently
used audio device in the UI the renderer needs to report what device is
in use. Currently, MediaWebContentsObserver will receive messages when
the output device is switched. However, the device id reported by the
renderer is hashed so that it may be only used on one origin.

This change has the MediaWebContentsObserver translate the device ID
before passing it off to the MediaSessionControllersManager.

Bug: 1111432
Change-Id: I11d161ff5f89525b4c465f01a779bdd30e31c823
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2331330
Commit-Queue: Noah Rose Ledesma <noahrose@google.com>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795175}