PM: Add HasObserver method to V8PerFrameMemoryRequestAnySeq.
For an Ad metrics project, we would like to gate adding an observer to
a subclass of V8PerFrameMemoryRequestAnySeq on there being an ad frame
detected on the page. Hence, we need to be able to check whether we
added this observer so that we can safely remove it in the destructor.
This CL adds a public bool method for checking whether the observer is
there before we try to remove it.

Change-Id: I2a212eb91edbb81e952d629b364652937e93c995
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327814
Reviewed-by: Joe Mason <joenotcharles@chromium.org>
Reviewed-by: Chris Hamilton <chrisha@chromium.org>
Commit-Queue: Chris Hamilton <chrisha@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795220}