Clean up appearance only picker tests
Create a new API to documents picker tests that only open
the picker to validate appearance.

Also fixing a bug in the color-picker-appearance-disabled.html test. For
a disabled color control, it is expected that the popup will not open.
Therefore, we should fail the test if the popup does open.

Bug: 1111999
Change-Id: Ica4ee5ce4db46d429e3949ec53a15b52e4ea5fc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333342
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Sanket Joshi <sajos@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#795258}