cros: Close wallpaper preview when overview mode is toggled.
Currently if a user is previewing a wallpaper and they enter overview
mode it causes a visual glitch where the overview item for the wallpaper
picker is transparent.

This CL makes it so Ash can emit a new event through the
chrome.wallpaper_private API, onClosePreviewWallpaper, to tell the app
side of wallpaper preview to close its front-end. This CL also adds
the necessary handling for the new event to the wallpaper manager.

Bug: 895265
Change-Id: I127fcedcfbd8c19564043910b3eac793dac46995
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2329414
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Commit-Queue: Jeremy Chinsen <chinsenj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795263}