[filesapp] Default to installed apps over view-in-browser fallback
When users have no default app for a file type we choose the application
task that appears as the default from a linearly ordered set that's
constructed in (sort of) arbitrary order.

Add an extra pass over candidate matches that checks if an installed
application was included as a fallback before we check for any
view-in-browser fallback.

       --gtest_filter="*ChooseAndSetDefaultTask_FallbackTextApp"

Bug: 1111399, 1112406
Tests: unit_tests
Change-Id: I328b79e3ababe2752bd7c1f8930568ea5f696d4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332079
Reviewed-by: Bugs Nash <bugsnash@chromium.org>
Commit-Queue: Alex Danilo <adanilo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795273}