Fix bug in "Add to Existing Tab Group"
This regression was caused by this refactor:
https://chromium-review.googlesource.com/c/chromium/src/+/2324113

The original code for building the submenu would skip ID values for
groups that would not appear in the submenu. (The Build() method would
call GetOrderedTabGroups then skip over groups that should not show in
the submenu.) Skipping ID values was not done after the refactor since
the subclasses would call the base classes with a vector of MenuItemInfo
to populate the submenu.

Bug: 1112719, 1112721, 1108062
Change-Id: I4a3490d646c553d1c60dd548fa13f53df9183561
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337547
Reviewed-by: Charlene Yan <cyan@chromium.org>
Reviewed-by: Connie Wan <connily@chromium.org>
Commit-Queue: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795293}