[b4p] Suppress move bubble for the primary account password
If the password of the primary signed-in account is stored locally (i.e.
in the profile store), the bubble that offers moving it into the account
should not be shown, since storing the password for an account within
that same account doesn't make sense. Moving other Google account
passwords is still valid.

Bug: 1110894
Change-Id: I2be5b30ecb121e5db91e3ff33b98cba51af6430b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339980
Commit-Queue: Victor Vianna <victorvianna@google.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795438}