[Nearby] Add settings behavior and stub out on-boarding
This is a mock up of the Nearby Share on-boarding page. Specs/css
strings, accessibility to be refined in another CL. This UI is gated
behind the feature flag.

Screenshot: https://screenshot.googleplex.com/hN3F6gjGFBJ

The settings behavior wraps up the mojo interface into a Polymer
property which can be passed down from the root object to the children
to allow them to bind and edit settings. In this case, the app.js
element hosts the settings object and passed it down to the on-boarding
page with two-way binding.

Change-Id: I2f1dbe25d0706597c94fc80393084570610424f1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332988
Reviewed-by: Richard Knoll <knollr@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: James Vecore <vecore@google.com>
Cr-Commit-Position: refs/heads/master@{#795472}