Introduce an allowlist for external experiment ids.
The allowlist provides a way to specify which external experiment
ids can be sent and also provides a way to associate a study name
for such external experiment ids.

Since the API now uses an allowlist, its use is no longer restricted
via metrics service accessor, which allows us to simplify the plumbing.

A metrics client API is provided to disable the allowlist, which is
needed for some configurations such as WebLayer.

Additionally, the new allowlist support is implemented behind a
base::Feature, which is off by default in this CL. When the
feature is disabled, the behavior from before is unchanged.

Bug: 1111941
Change-Id: I63b92254e27a32dd5236c37f0b70fb3683f447e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2324172
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Peter Conn <peconn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795502}