IdleWakeups: refactor SystemInformationSampler::TakeSnapshot
This change makes SystemInformationSampler::TakeSnapshot() smaller by
extracting a block of code into a separate function, shortening this
long function from 88 lines to a slightly more readable 58 lines.

This change also moves an early return for validation to the earliest
place it can be, slightly clarifies a comment, and fixes a typo in
IdleWakeups.

Change-Id: I3ef0271e82de74b87d1b9ce57b4bd38dc834bd23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2335797
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Cr-Commit-Position: refs/heads/master@{#795522}