Make profile optional when searching for Tab Groups across browsers
Followup to crrev.com/c/2326632 and some comments there.

For the Tab Groups Extensions API, it's useful to be able to query all browsers for the one containing a TabGroupId, without needing to know the profile up-front. There are separate checks for ensuring that a group doesn't move across profiles. This is still assuming it's unlikely for group IDs to clash across profiles.

If there's a better way to approach this I'm happy to adjust!

Bug: 1098439
Change-Id: I71bd07f5d00fad74ee5aa02d1a62feb4883e3819
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339341
Commit-Queue: Connie Wan <connily@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795549}