Set NameFrom property for text nodes in PdfAccessibilityTree
This CL sets the ax::mojom::NameFrom property for static text nodes and
inline text nodes in PdfAccessibilityTree. This property is used by
AXNode::GetInnerText() to compute the inner text of the nodes.

This CL also includes the relevant tests to validate the change.

Bug: 1112890
Change-Id: I782ceb4d93a07146a81b1d23ff9696af3267d7f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2337195
Commit-Queue: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Ankit Kumar 🌪️ <ankk@microsoft.com>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795564}