device/fido: turn the caBLEv2 handshake around.
The current caBLEv2 handshake has the desktop speak first. We expect the
network-based design to have the phone speak first, thus this change
turns the handshake “around” to accommodate that. Also, the current v2
handshake includes pairing information in the response but, if the phone
speaks first, the response is coming from the wrong place to include
pairing information and the first message has insufficient protection to
include it. Thus that is removed. Instead, a getInfo message can be
included in the phone's first message because it'll be needed anyway and
it's generic and thus doesn't require forward security.

BUG=1002262

Change-Id: If2fb27b2a56ef21c0b21a454ff498f65ce62e81f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2340062
Commit-Queue: Adam Langley <agl@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#795679}