VideoToolbox/SharedImage: Plumb SharedImageStub
Add a method through which VTVideoDecodeAccelerator will be able to
access a SharedImageStub. In this case we've hung it off of
VdaVideoDecoder's implementation of the VideoDecodeAccelerator::Client
implementation. An alternative could be the GpuVideoDecodeGLClient, but
that's a GL thing and after this change, VTVideoDecodeAccelerator won't
use any GL things.

Bug: 1108909
Change-Id: I46e7e92eaa90e7fb89c4582e53d985c3ae1d8d4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2333304
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795706}