Add ukm::SourceIds to ServiceWorkers
This cl adds a ukm::SourceId to each ServiceWorkerVersion and propagates
that down to the ServiceWorkerGlobalScope.

A followup cl will have each service worker client inform its respective
ServiceWorkerContainerHost of the client's ukm::SourceId. A UKM event
will also be emitted on on that connection in order to associate the two
ukm::SourceIds.

Bug: 1085645
Change-Id: I9a77e4196834379162c8e7516eb0bde58f5fa5e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298349
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795710}