Make DocumentInit use LocalDOMWindow instead of LocalFrame/DocumentLoader
This only matters for Documents that will be attached, but given that
the Document will attach to the LocalDOMWindow, that should be the
primary object the DocumentInit holds.

Also, while cleaning up state on DocumentInit, remove GetUseCounter().
It's only use is during the Document constructor, and we have another
reliable path to the UseCounter.

Change-Id: Iafbb3c1c335c37bdf5bcf7e2ed3856ebf22bc861
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339399
Auto-Submit: Nate Chapin <japhet@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795716}