overview: Ensure auto snapping from overveiw for async apps in tablet splitview
This CL fixes the issue that selecting async apps (e.g. ARC apps) from overview does not trigger auto snapping in tablet splitview.
In tablet splitview, when a window is snapped, we can select which window to fill (be snapped in) the other side.
It works for synchronized Chrome windows, but not for async windows because calling Unminimize(), which is used for canceling the restore animation, bypasses auto snapping.
So we should use Show() here and extract async activation logic from OverviewItem to make sure it works for async apps.

BUG=b:162552497
TEST=SelectMinimizedSnappableWindowInSplitView

Change-Id: I653451533c58a6241130a5ab369bb080290810af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2327364
Commit-Queue: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795727}