Extensions: Add permissions for metricsPrivate to Tab Search WebUI
The Chrome for ChromeOS team is looking to collect UMA metrics for
the time taken to render the initial list of tabs in our Tab Search
feature's UI.

This metric needs to be logged from the JS side and access to the
Extensions API provides a clean way to log this data. This is the
recommended way to log metrics in WebUI code.

Bug: 1099917
Change-Id: Icf16404d47406858fd3c6cef2ba4b8462f2edb18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339211
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795736}