Support reverting web app migration if migration flag is disabled
This CL ensures we re-install default installed Extension apps that
were migrated to web apps if the migration flag is disabled.

Previously we would not reinstall the old Extension because the
uninstall for migration was recorded as a user uninstall.

This CL updates AppService.Uninstall() to take an |uninstall_source|
enum so the call site can indicate whether it is a user or migration
uninstall.

Bug: 1104696
Change-Id: I832e2d8f3c3769a45f147aafb1c4e15646c8fdae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2339294
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Glen Robertson <glenrob@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795782}