CrOS: Properly size and scale icons when standardizing with a background
This CL creates a background circle of the proper size for standardized
icons.

Also, icons that should appear within a background circle are now
scaled by using the distance of the farthest visible point from the
center of the icon. This will ensure a constant spacing between the
scaled down icons and the background circle.

Bug: 1110496
Change-Id: I392b9a3738f2dac01dadea44b8a3e473402298eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342066
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Matthew Mourgos <mmourgos@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795808}