Modify the mask image loading to generate the correct scaled image.
When the scale is not 1.0, e.g. 2.0, the web apps generated icon can't
apply the mask correctly, because the web apps icons use the scaled
ImageSkiaReps, which don't match the mask imageskia loaded from the
ResourceBundle, and CreateMaskedImage reports error due to mismatched
scale and pixel size.

Add a mask cache to cache the mask ImageSkiaReps for different scale,
pixel size, and when loading the mask, check the current image's scale,
pixel size to generate the mask imageskia to match the image.

BUG=1083331

Change-Id: If16c3603db888e73bb272e6cbdd409ba8d0d4468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2336387
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795811}