[Android] Add SyncOffPreference in ManageSyncSettings page
This cl adds the SyncOffPreference in the ManageSyncSettings page. It
also adds "Advanced" category after the Preference. The preference is
only shown after the user has signed in. So it is not shown in the
advanced sync consent flow.
The code for ClearDataProgressDialog from AccountManagementFragment
is reused here.

Bug: 1092410
Change-Id: I0b70ad5cfc32d4c8325680f9929aaae6ccfca65f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273193
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795840}