[Autofill Assistant] Remove intent validation for password scripts
As password scripts have a confirmation step, we don't need to validate
an intent to change a password.
We still need to check whether a running script is a password script (if
yes, the password manager should be notified that a script is running).
For that, we will check whether |TriggerContext| has
PASSWORD_CHANGE_USERNAME param. (For convenience, this CL moves that
parameter's name and also kOverlayColorsParameterName to
trigger_context.cc).

This CL also changes what origin is used for password store request. If
a password script is started from Settings, GetLastCommittedURL is not
set yet. So, we have to use the deep url provided.

Bug: 1086109
Change-Id: I7c38388b06f433dc51bce1a2b54a7d48bb322bfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332594
Commit-Queue: Maxim Kolosovskiy  <kolos@chromium.org>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Milica Selakovic <selakovic@google.com>
Cr-Commit-Position: refs/heads/master@{#795844}