[blink/bindings] Make ParkableStringsToDisk depend on compression.
ParkableStringsToDisk depends on compression being enabled. This is not
an issue in most cases, since compression is enabled by default. It is
however disabled in some environments, such as some headless cases. If
the two features are not tied in the code, this will cause issues once
ParkableStringsToDisk is enabled by default.

Bug: 1029320
Change-Id: I78a1053d54dcd4103f3017cd512428753b75e537
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342849
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Benoit L <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#795875}