HaTS Next: Add web dialog to deliver web based survey
To support delivery of HaTS Next surveys, a new web dialog is created
which connects to the Hats Next Chrome specific website. The dialog
responds to events provided by the survey for managing dialog lifetime.

For example the survey dialog is only shown when the HaTS webpage has
confirmed that the user is eligible for a survey.

Bug: 1110888
Change-Id: I323c662e41331cab47f05424de53906169b42f23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246695
Reviewed-by: Theodore Olsauskas-Warren <sauski@chromium.org>
Reviewed-by: Bret Sepulveda <bsep@chromium.org>
Commit-Queue: Theodore Olsauskas-Warren <sauski@google.com>
Cr-Commit-Position: refs/heads/master@{#795911}